\name{dikarya}
\alias{dikarya}
\title{
Tree of Dikarya
}
\description{Phylogenetic tree of true fungi (Dikarya)}
\usage{data("dikarya")}

\details{The topology is based on published articles. See 'References' for the article list.}
\value{
\item{dikarya}{object of class \code{phylo}}
}
\references{
Floudas et al. (2012) \emph{The Paleozoic origin of enzymatic lignin decomposition reconstructed from 31 fungal genomes.} Science 336:1715-1719

Padamsee et al. (2012) \emph{The genome of the xerotolerant mold Wallemia sebi reveals adaptations to osmotic stress and suggests cryptic sexual reproduction.} Fung Genet Biol 49:217-226

Matheny et al. (2006) \emph{Major clades of Agaricales: a multilocus phylogenetic overview.} Mycologia 98:982-995
}
\author{
Franz-Sebastian Krah
}
\examples{
data(dikarya)
}