\name{reference.list}
\alias{reference.list}
\title{
Lockup complete reference list from TIME TREE
}
\description{
Lockup divergence time from TIME TREE}
\usage{
reference.list()
}
\details{Downloads the current reference list for the TIMETREE OF LIFE. It serves information on Author, Title, Journal, Year and PubMed ID}
\value{
an object of class list
\item{}{a data.frame with reference information}
}
\author{
Franz-Sebastian Krah
}
\examples{
# reference.list() }