% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetree_new.R
\name{timetree}
\alias{timetree}
\title{Retrieve records from the timetree website}
\usage{
timetree(taxon1 = "Agaricus", taxon2 = "Boletus", verbose = TRUE,
  screenshot = TRUE, port = 4445L, browserName = "chrome",
  remoteServerAddr = "localhost", wait = 2)
}
\arguments{
\item{taxon1}{character string specifying the first taxon name}

\item{taxon2}{character string specifying the second taxon name}
}
\description{
Retrieve records from the timetree website
}
\examples{
\dontrun{
timetree(taxon1 = "Agaricus", taxon2 = "Boletus")
}
}
\author{
Franz-Sebastian Krah
}
